Checking if pCloud isn't installed before try to install him

+++ plugins/pcloud-installer.sh
@@ -2,14 +2,19 @@
 # Script written by: Jack Miras.
 # This script install pCloud.
 
-# Downloading pCloud to the Download folder.
-wget -P $HOME/Downloads/ https://c75.pcloud.com/dHZ8IsmSZJvU4uZZZmMutt7ZHkZZGNRZkZ5LUXZssw368uiJLjYQDTAP0KEsS2b0egy/pCloud_Linux_amd64_3.1.1.deb
+if command -v psyncgui >/dev/null; then
+  echo 'pCloud is already installed.'
+else
+  # Downloading pCloud to the Download folder.
+  wget -P $HOME/Downloads/ https://c132.pcloud.com/dHZ8IsmSZJvU4uZZZn1ij67ZHkZZHaRZkZ5LUXZ1nYdW4mJmNVDaaGfw8OCvLtCBaP7/pCloud_Linux_amd64_3.1.1.deb
 
-# Installing pCloud.
-sudo dpkg -i $HOME/Downloads/pCloud_Linux_amd64_3.1.1.deb && sudo apt-get -f install
+  # Installing pCloud.
+  sudo dpkg -i $HOME/Downloads/pCloud_Linux_amd64_3.1.1.deb && sudo apt-get -f install
 
-# Updating Ubuntu dependencies and softwares.
-sudo apt-get update && sudo apt-get upgrade
+  # Updating Ubuntu dependencies and softwares.
+  sudo apt-get update && sudo apt-get upgrade
 
-# Removing pCloud_Linux_amd64_3.1.1.deb file from Download folder.
-rm -rf $HOME/Downloads/pCloud_Linux_amd64_3.1.1.deb
+  # Removing pCloud_Linux_amd64_3.1.1.deb file from Download folder.
+  rm -rf $HOME/Downloads/pCloud_Linux_amd64_3.1.1.deb
+fi
+
